# openapi_export.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SME ERP API",
    "description": "Enterprise Resource Planning for Small and Medium Enterprises",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/inventory/reports/snapshot": {
      "get": {
        "tags": [
          "Inventory Reports"
        ],
        "summary": "Inventory snapshot (VIEWER+)",
        "description": "Get current inventory snapshot with filtering and pagination.\nRequires VIEWER role or higher.\n\nThis endpoint provides a comprehensive view of current stock levels across all locations.",
        "operationId": "get_inventory_snapshot_api_v1_inventory_reports_snapshot_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by location ID",
              "title": "Location Id"
            },
            "description": "Filter by location ID"
          },
          {
            "name": "item_sku",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item SKU",
              "title": "Item Sku"
            },
            "description": "Filter by item SKU"
          },
          {
            "name": "item_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search item name (case-insensitive)",
              "title": "Item Name"
            },
            "description": "Search item name (case-insensitive)"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item status",
              "title": "Status"
            },
            "description": "Filter by item status"
          },
          {
            "name": "min_quantity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum stock quantity",
              "title": "Min Quantity"
            },
            "description": "Minimum stock quantity"
          },
          {
            "name": "max_quantity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum stock quantity",
              "title": "Max Quantity"
            },
            "description": "Maximum stock quantity"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Number of records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrentStockOut"
                  },
                  "title": "Response Get Inventory Snapshot Api V1 Inventory Reports Snapshot Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/reports/movements": {
      "get": {
        "tags": [
          "Inventory Reports"
        ],
        "summary": "Stock movement history (VIEWER+)",
        "description": "Get stock movement history with comprehensive filtering.\nRequires VIEWER role or higher.\n\nProvides detailed audit trail of all inventory movements.",
        "operationId": "get_stock_movements_api_v1_inventory_reports_movements_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item ID",
              "title": "Item Id"
            },
            "description": "Filter by item ID"
          },
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by location ID",
              "title": "Location Id"
            },
            "description": "Filter by location ID"
          },
          {
            "name": "transaction_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by transaction type",
              "title": "Transaction Type"
            },
            "description": "Filter by transaction type"
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "From date (YYYY-MM-DD)",
              "title": "From Date"
            },
            "description": "From date (YYYY-MM-DD)"
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "To date (YYYY-MM-DD)",
              "title": "To Date"
            },
            "description": "To date (YYYY-MM-DD)"
          },
          {
            "name": "reference_no",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by reference number",
              "title": "Reference No"
            },
            "description": "Filter by reference number"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Number of records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockLedgerOut"
                  },
                  "title": "Response Get Stock Movements Api V1 Inventory Reports Movements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/reports/snapshot/csv": {
      "get": {
        "tags": [
          "Inventory Reports"
        ],
        "summary": "Export inventory snapshot to CSV (ADMIN+)",
        "description": "Export inventory snapshot to CSV format.\nRequires ADMIN role or higher.\n\nWARNING: This endpoint can generate large files. Use filters to limit data size.",
        "operationId": "export_inventory_snapshot_csv_api_v1_inventory_reports_snapshot_csv_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by location ID",
              "title": "Location Id"
            },
            "description": "Filter by location ID"
          },
          {
            "name": "item_sku",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item SKU",
              "title": "Item Sku"
            },
            "description": "Filter by item SKU"
          },
          {
            "name": "item_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search item name",
              "title": "Item Name"
            },
            "description": "Search item name"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by item status",
              "title": "Status"
            },
            "description": "Filter by item status"
          },
          {
            "name": "min_quantity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum stock quantity",
              "title": "Min Quantity"
            },
            "description": "Minimum stock quantity"
          },
          {
            "name": "max_quantity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum stock quantity",
              "title": "Max Quantity"
            },
            "description": "Maximum stock quantity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/inventory/reports/summary": {
      "get": {
        "tags": [
          "Inventory Reports"
        ],
        "summary": "Inventory summary statistics (VIEWER+)",
        "description": "Get high-level inventory summary statistics.\nRequires VIEWER role or higher.\n\nProvides key metrics for dashboard and reporting.",
        "operationId": "get_inventory_summary_api_v1_inventory_reports_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Inventory Summary Api V1 Inventory Reports Summary Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrentStockOut": {
        "properties": {
          "item_id": {
            "type": "integer",
            "title": "Item Id"
          },
          "item_sku": {
            "type": "string",
            "title": "Item Sku"
          },
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "location_id": {
            "type": "integer",
            "title": "Location Id"
          },
          "location_code": {
            "type": "string",
            "title": "Location Code"
          },
          "location_name": {
            "type": "string",
            "title": "Location Name"
          },
          "current_quantity": {
            "type": "string",
            "title": "Current Quantity"
          },
          "last_transaction_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Transaction Date"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "item_sku",
          "item_name",
          "location_id",
          "location_code",
          "location_name",
          "current_quantity",
          "last_transaction_date"
        ],
        "title": "CurrentStockOut",
        "description": "Current stock summary by item and location"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}